fix(static): path traversal handling on windows
comment(@casperx): The Path parses input strings in a platform-specific way. On Windows, it converts the forward slash ("/") into a backslash ("\"). However, when using Path to parse a URL path, which always uses the forward slash as a component separator, this conversion can lead to issues. Specifically, when compiled on Windows, this can result in backslashes appearing in path components.

We could simply add the backslash to the path segment handler condition to be handled the same way as a forward slash. However, this solution is not platform-independent, as a backslash might be a valid directory name on a non-Windows platform. Instead, we should use the `RelativePath` to parse URL paths, as it consistently uses the forward slash. Additionally, we can simplify our path component matching to just "." and ".." because `RelativePath` does not emit path separators like the `Path` does.